Intro: fill in "open source and more info" section
Partially fulfills #74

This section is a little bit different than the one in #81 and #83,
since the short blurb does not have any links, and each of the
highlights are instead the links. We will also probably add more links
to this section as we create more content related to this topic.

I linked to just a coupl things to start off, and used the same
placeholder SVG that I created for #83. I'm not sure if we want to
create some sort of iconography for each of these links, or if we should
use something like the YouTube placeholder for the video, or what.