Intro: use Footer--column class for performance
Using the `> *` selector in the `Columns` component makes sense, since
it accepts any children, and turns them into columns. However, it
isn't as necessary here, since the HTML and the CSS are closely coupled
anyhow. It's easy enough to use a specific class name rather than using
a universal selector.