Intro: style more like citrusbyte.com footer
Now that I have it for reference (I am not on a plane anymore, and can
get to citrusbyte.com), I can see that it does many things better than
what I was doing. This takes the parts that work with the flow of this
particular page, but does not fully copy it.