Intro: improve gif-to-earth progress bar
Make it look more like the one in #80

I made a judgement call to leave the gif below the bar, instead of
putting it on top. This is because I don't know what the style should
look like when the image is right next to the planets. The only good way
to do that I can think of seems like it will take too long to code.

Also, the designs show the arrow being aligned to the bottom of the
page. To accomplish this would require restructuring a big portion of
the app, and it doesn't seem worth it to me. So I also made a judgement
call to leave this as-is.