Intro: Simplify "GIF Arrived" section
Use styles provided by Sam. No Call To Action to download or share the
GIF they created, instead the only CTA is to visit Thicket.

This is partly because it was taking monstrous long to try to get even
just GIF-downloading to work. Let alone sharing the GIF, which would
require importing `js-ipfs` and recreating Thicket here on the intro
page.

It's also nice that there's only one CTA, rather than two.