Intro: clear data after 24 hours
We want to eventually reset the site how it was. Sam and I discussed and
thought it made sense to reset everything automatically, with no UI
element to clear it manually. That way it's nice and focused, and
hopefully gets people to click over to Thicket faster.

We decided that it makes sense to do this automatic reset after one day.